requirements.txt
django
djangorestframework
pillow
djangorestframework_simplejwt
requests
telebot
psutil
email-validator
